move interface location to changeNetwork file.
