+++ src/lib/provider.ts
@@ -1,3 +1,5 @@
+import { AddEthereumChainParameter } from '../ux/wrongNetwork/changeNetwork'
+
 // https://eips.ethereum.org/EIPS/eip-1193
 interface RequestArguments {
   readonly method: string;
@@ -8,19 +10,6 @@ export interface EIP1193Provider {
   request<T = string>(args: RequestArguments): Promise<T>
   isMetaMask: boolean | null
   isNiftyWallet: boolean | null
-}
-
-export interface AddEthereumChainParameter {
-  chainId: string;
-  chainName: string;
-  nativeCurrency: {
-    name: string;
-    symbol: string;
-    decimals: 18;
-  };
-  rpcUrls: string[];
-  blockExplorerUrls?: string[];
-  iconUrls?: string[]; // Currently ignored.
 }
 
 export const ethAccounts = (provider: EIP1193Provider) => provider.request<string[]>({ method: 'eth_accounts' })
